refactor: 병합후 파일경로 변경
